create order

+++ db/migrate/20171129134048_create_carts.rb
@@ -2,7 +2,7 @@ class CreateCarts < ActiveRecord::Migration[5.1]
   def change
     create_table :carts do |t|
       t.references :user, foreign_key: true
-      t.integer :total_money
+      t.float :total_money
       t.references :food, foreign_key:true
       t.integer :quantity
